requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
entrypoints==0.4
fastapi==0.95.1
ffmpeg==1.4
ffmpeg-downloader==0.2.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
ffprobe==0.5
ffprobe-python==1.0.3
filelock==3.12.0
Flask==2.2.2
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
google-api-core==2.11.0
google-auth==2.18.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gradio==3.27.0
gradio_client==0.1.3
gunicorn==20.1.0
h11==0.14.0
heyoo==0.0.8
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
orjson==3.8.10
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
plotly==5.14.1
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rich==13.3.4
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
starlette==0.26.1
tenacity==8.2.2
threadpoolctl==3.1.0
toolz==0.12.0
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0.2
Werkzeug==2.2.3
yarl==1.8.2
